Rework the contract between controller and daemons
The new processors have been integrated to handle the different
connection times

Also renamed BATCHD_CONNECTED to STEPD_CONNECTED
