+++ lib/flight_scheduler/event_processor.rb
@@ -26,213 +26,6 @@
 #==============================================================================
 
 module FlightScheduler::EventProcessor
-  module ProcessorHelper
-    def read_loop
-      while message = connection.read
-        process(message)
-      end
-      connection.close
-    end
-
-    # TODO: Make private
-    def process(message)
-      raise NotImplementedError
-    end
-  end
-
-  DaemonProcessor = Struct.new(:connection, :node_name) do
-    include ProcessorHelper
-
-    def process(message)
-      case message[:command]
-      when 'NODE_DEALLOCATED'
-        node_deallocated(message[:job_id])
-      end
-    end
-
-    def send_job_allocated(job_id, environment:, username:, time_limit:)
-      connection.write(
-        command: 'JOB_ALLOCATED',
-        job_id: job_id,
-        environment: environment,
-        username: username,
-        time_limit: time_limit
-      )
-      connection.flush
-      Async.logger.info("Sent JOB_ALLOCATED to #{node_name} (job: #{job_id})")
-    end
-
-    # NOTE: This maybe better suited on JobProcessor but this would require changing
-    # the relationship between daemon-jobd/batchd to:
-    # 1. Send message to jobd/batchd to trigger a graceful shutdown,
-    # 2. The daemon detects the shutdown and responds NODE_DEALLOCATED implicitly
-    # X. What happens if it is in a state that can't be shutdown? (e.g. steps running)
-    def send_job_deallocated(job_id)
-      connection.write(
-        command: 'JOB_DEALLOCATED',
-        job_id: job_id
-      )
-      connection.flush
-      Async.logger.info("Sent JOB_DEALLOCATED to #{node_name} (job: #{job_id})")
-    end
-
-    private
-
-    def node_deallocated(job_id)
-      # Remove the node from the job
-      # The job's allocation will be remove implicitly if this was the last node
-      job = FlightScheduler.app.job_registry[job_id]
-      allocation = FlightScheduler.app.allocations
-                                  .deallocate_node_from_job(job_id, node_name)
-      return unless allocation
-
-      if allocation.nodes.empty? && !job.terminal_state?
-        # If the job is not in a terminal state, it has not been updated
-        # following a `NODE_{COMPLETED,FAILED}_JOB` command.  It might have been
-        # created without a batch script, i.e., created with the `alloc`
-        # command.  Or the message might not have been sent or received.
-        # Ideally, we'd capture the exit code of some command somewhere to be
-        # able to set the FAILED state if appropriate.
-        if job.state == 'CANCELLING'
-          job.state = 'CANCELLED'
-        elsif job.state == 'TIMINGOUT'
-          job.state = 'TIMEOUT'
-        else
-          job.state = 'COMPLETED'
-        end
-      end
-      FlightScheduler.app.processors.event.allocate_resources_and_run_jobs
-    end
-  end
-
-  JobProcessor = Struct.new(:connection, :node_name, :job_id) do
-    include ProcessorHelper
-
-    def process(message)
-      case message[:command]
-      when 'NODE_COMPLETED_JOB'
-        node_completed_job
-      when 'NODE_FAILED_JOB'
-        node_failed_job
-      when 'JOB_TIMED_OUT'
-        job_timed_out
-      end
-    end
-
-    def send_run_script(arguments:, script:,  stdout_path:, stderr_path:)
-      connection.write({
-        command: 'RUN_SCRIPT',
-        arguments: arguments,
-        job_id: job_id,
-        script: script,
-        stdout_path: stdout_path,
-        stderr_path: stderr_path
-      })
-      connection.flush
-      Async.logger.info("Sent RUN_SCRIPT to #{node_name} (job: #{job_id})")
-    end
-
-    def send_run_step(arguments:, path:, pty:, step_id:, environment:)
-      connection.write({
-        command: 'RUN_STEP',
-        arguments: arguments,
-        job_id: job_id,
-        path: path,
-        pty: pty,
-        step_id: step_id,
-        environment: environment
-      })
-      connection.flush
-      Async.logger.info("Sent RUN_STEP to #{node_name} (job: #{job_id} - step: #{step_id})")
-    end
-
-    def send_job_cancelled
-      connection.write({
-        command: 'JOB_CANCELLED',
-        job_id: job_id
-      })
-      connection.flush
-      Async.logger.info("Sent JOB_CANCELLED to #{node_name} (job: #{job_id})")
-    end
-
-    private
-
-    def node_completed_job
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      return if job.nil?
-
-      Async.logger.info("Node #{node_name} completed job #{job.display_id}")
-      job.state = 'COMPLETED'
-      FlightScheduler::Deallocation::Job.new(job).call
-    end
-
-    def node_failed_job
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      return if job.nil?
-
-      Async.logger.info("Node #{node_name} failed job #{job.display_id}")
-      if job.state == 'CANCELLING'
-        job.state = 'CANCELLED'
-      elsif job.state == 'TIMINGOUT'
-        job.state = 'TIMEOUT'
-      else
-        job.state = 'FAILED'
-      end
-      FlightScheduler::Deallocation::Job.new(job).call
-    end
-
-    def job_timed_out
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      return unless job
-      job.state = job.terminal_state? ? 'TIMEOUT' : 'TIMINGOUT'
-    end
-  end
-
-  StepProcessor = Struct.new(:connection, :node_name, :job_id, :step_id) do
-    include ProcessorHelper
-
-    def process(message)
-      case message[:command]
-      when 'RUN_STEP_STARTED'
-        job_step_started(message[:port])
-      when 'RUN_STEP_COMPLETED'
-        job_step_completed
-      when 'RUN_STEP_FAILED'
-        job_step_failed
-      end
-    end
-
-    private
-
-    def job_step_started(port)
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      job_step = job.job_steps.detect { |step| step.id == step_id }
-      Async.logger.info("Node #{node_name} started step #{job_step.display_id}")
-      execution = job_step.execution_for(node_name)
-      execution.state = 'RUNNING'
-      execution.port = port
-      FlightScheduler.app.persist_scheduler_state
-    end
-
-    def job_step_completed
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      job_step = job.job_steps.detect { |step| step.id == step_id }
-      Async.logger.info("Node #{node_name} completed step #{job_step.display_id}")
-      execution = job_step.execution_for(node_name)
-      execution.state = 'COMPLETED'
-      FlightScheduler.app.persist_scheduler_state
-    end
-
-    def job_step_failed(node_name, job_id, step_id)
-      job = FlightScheduler.app.job_registry.lookup(job_id)
-      job_step = job.job_steps.detect { |step| step.id == step_id }
-      Async.logger.info("Node #{node_name} failed step #{job_step.display_id}")
-      execution = job_step.execution_for(node_name)
-      execution.state = 'FAILED'
-      FlightScheduler.app.persist_scheduler_state
-    end
-  end
-
   def job_created(job)
     FlightScheduler.app.job_registry.add(job)
     allocate_resources_and_run_jobs
@@ -244,11 +37,6 @@ module FlightScheduler::EventProcessor
     FlightScheduler::Submission::JobStep.new(job_step).call
   end
   module_function :job_step_created
-
-  def node_connected
-    allocate_resources_and_run_jobs
-  end
-  module_function :node_connected
 
   def allocate_resources_and_run_jobs
     Async.logger.info("Attempting to allocate rescources to jobs")
